@@ -2,7 +2,7 @@
   "name": "@tolgee/core",
   "version": "1.0.0-alpha.60",
   "description": "Library providing ability to translate messages directly in context of developed application.",
-  "main": "dist/tolgee.umd.js",
+  "main": "index.js",
   "types": "./lib/index.d.ts",
   "repository": {
     "type": "git",
@@ -22,45 +22,34 @@
     "dist": "dist"
   },
   "scripts": {
-    "build": "webpack --env mode=production && tsc",
+    "build": "rollup -c rollup.config.js && tsc",
     "test": "jest --config ./jest.config.js",
-    "develop": "webpack --env mode=development --watch"
+    "develop": "concurrently \"rollup -c rollup.config.js -w\" \"tsc --watch\" "
   },
   "author": "JanCizmar",
   "license": "ISC",
   "files": [
+    "index.js",
     "lib/**/*",
-    "dist/**/*",
-    "core/**/*",
-    "ui/**/*"
+    "dist/**/*"
   ],
   "devDependencies": {
+    "@rollup/plugin-node-resolve": "13.0.4",
+    "@rollup/plugin-typescript": "8.2.5",
     "@testing-library/dom": "^7.24.5",
     "@testing-library/jest-dom": "^5.11.4",
     "@types/intl-messageformat": "^3.0.0",
     "@types/jest": "^26.0.14",
     "@types/node": "^14.14.22",
-    "clsx": "^1.1.1",
-    "copy-webpack-plugin": "^7.0.0",
-    "css-loader": "^5.0.1",
-    "cypress": "^7.1.0",
-    "cypress-dotenv": "^1.2.2",
-    "cypress-xpath": "^1.6.1",
-    "dotenv-webpack": "^7.0.2",
-    "html-webpack-plugin": "^4.5.0",
-    "intl-messageformat": "^9.5.3",
+    "concurrently": "6.2.1",
+    "intl-messageformat": "^9.9.1",
     "jest": "^26.5.2",
     "jest-fetch-mock": "^3.0.3",
-    "reflect-metadata": "^0.1.13",
-    "regenerator-runtime": "^0.13.7",
+    "rollup": "^2.56.3",
+    "rollup-plugin-terser": "7.0.2",
     "ts-jest": "^26.4.1",
-    "ts-loader": "^8.0.14",
-    "tslint": "~6.1.3",
-    "tsyringe": "^4.3.0",
-    "typescript": "^4.0.5",
-    "webpack": "^5.21.2",
-    "webpack-bundle-analyzer": "^4.3.0",
-    "webpack-cli": "^4.5.0"
+    "tslib": "^2.3.1",
+    "typescript": "4.3.5"
   },
   "gitHead": "d4f5dccf1f3b0017942d862f7be7e9efe093ef08",
   "publishConfig": {
