@@ -2,21 +2,20 @@ jest.dontMock('./AttributeHandler');
 jest.dontMock('./AbstractHandler');
 jest.dontMock('../services/EventService');
 jest.dontMock('../helpers/NodeHelper.ts');
+jest.dontMock('../services/DependencyStore');
 
+import { Properties } from '../Properties';
+import { AttributeHandler } from './AttributeHandler';
 import { ElementWithMeta, NodeWithMeta } from '../types';
 import { NodeHelper } from '../helpers/NodeHelper';
-import describeClassFromContainer from '@testFixtures/describeClassFromContainer';
 import { createTestDom } from '@testFixtures/createTestDom';
 import { ReplacedType, TextService } from '../services/TextService';
 import { getMockedInstance } from '@testFixtures/mocked';
 import { ElementRegistrar } from '../services/ElementRegistrar';
+import { DependencyStore } from '../services/DependencyStore';
 
 describe('AttributeHandler', () => {
-  const getAttributeHandler = describeClassFromContainer(
-    import('./AttributeHandler'),
-    'AttributeHandler'
-  );
-  let attributeHandle: ReturnType<typeof getAttributeHandler>;
+  let attributeHandler: AttributeHandler;
 
   const mockedKeys = [
     {
@@ -38,21 +37,34 @@ describe('AttributeHandler', () => {
 
   beforeEach(() => {
     c = createTestDom(document);
-    attributeHandle = getAttributeHandler();
+    attributeHandler = new DependencyStore().attributeHandler;
+    getMockedInstance(Properties).config = {
+      inputPrefix: '{{',
+      inputSuffix: '}}',
+      restrictedElements: [],
+      tagAttributes: {
+        '*': ['aria-label'],
+      },
+      passToParent: ['option'],
+    };
     getMockedInstance(TextService).replace = async (...args) =>
       mockedTranslate(...args);
+  });
+
+  afterEach(async () => {
+    jest.clearAllMocks();
   });
 
   describe('in production mode', () => {
     let xPath: string;
 
     beforeEach(async () => {
-      await attributeHandle.handle(document.body);
+      await attributeHandler.handle(document.body);
       xPath = `./div/div/@aria-label[contains(., '${gv(c.ariaLabelKey)}')]`;
     });
 
     test('Can be created', () => {
-      expect(attributeHandle).not.toBeUndefined();
+      expect(attributeHandler).not.toBeUndefined();
     });
 
     test('will handle text in aria-label attribute of div', async () => {
@@ -90,7 +102,7 @@ describe('AttributeHandler', () => {
   });
 
   test('will register the node', async () => {
-    await attributeHandle.handle(document.body);
+    await attributeHandler.handle(document.body);
     const xPath = `./div/div/@aria-label[contains(., '${gv(c.ariaLabelKey)}')]`;
     const node = NodeHelper.evaluateToSingle(xPath, document.body) as Attr;
     expect(getMockedInstance(ElementRegistrar).register).toBeCalledWith(
