@@ -7,9 +7,12 @@ import { Page } from './Page';
 const App = () => {
   return (
     <TolgeeProvider
-      filesUrlPrefix="i18n/"
+      staticData={{
+        en: () => import('./i18n/en.json'),
+        de: () => import('./i18n/de.json'),
+      }}
       apiUrl={process.env.REACT_APP_TOLGEE_API_URL}
-      apiKey={process.env.REACT_APP_TOLGEE_API_KEY}
+      //apiKey={process.env.REACT_APP_TOLGEE_API_KEY}
       ui={UI}
       availableLanguages={['en', 'de']}
       loadingFallback={
